Rename: fuse_depth_images.cpp -> fuse_images.cpp
