+++ jsk_pcl_ros/include/jsk_pcl_ros/fuse_images.h
